+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This list is not currently intended to be all-encompassing - it will document major and breaking API changes with their rationale when appropriate:
 
+### v3.37.0
+- Upgrades to `http4k-testing-webdriver`. H/T @dickon for the PRs
+- Added `ProxyHost` request filter which is useful for writing proxy-type apps.
+
 ### v3.36.1
 - Fix #168 - Fix rest of hamkrest matchers caused by generics mishap.
 - Upgrade HTTP client dependency versions.